+++ test/cases/split-chunk-resolve-assets/src/assets/js/module-b.js
@@ -1,8 +1,0 @@
-const lib = require('./lib');
-const ansis = require('ansis');
-let value = lib.methodB();
-value = ansis.strip(value);
-
-console.log('>> common used module <<');
-
-module.exports = value;